Adding basic video functionality and openable project source
