+++ src/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=/home/jackjan/.netbeans/8.2/build.properties
