+++ src/src/com/hopding/jrpicam/enums/Encoding.java
