+++ src/src/com/hopding/jrpicam/enums/Exposure.java
