+++ src/src/com/hopding/jrpicam/enums/ImageEffect.java
