+++ src/src/com/hopding/jrpicam/enums/MeteringMode.java
